+++ tests/integration/hookregistration/HookRunner.php
@@ -1,0 +1,104 @@
+<?php
+
+namespace HookRegistration;
+
+interface Hook1Hook {
+	public function onHook1( $arg );
+}
+interface Hook2Hook {
+	public function onHook2( $arg );
+}
+interface Hook3Hook {
+	public function onHook3( $arg );
+}
+interface Hook4Hook {
+	public function onHook4( $arg );
+}
+interface Hook5Hook {
+	public function onHook5( $arg );
+}
+interface Hook6Hook {
+	public function onHook6( $arg );
+}
+interface Hook7Hook {
+	public function onHook7( $arg );
+}
+interface Hook8Hook {
+	public function onHook8( $arg );
+}
+interface Hook9Hook {
+	public function onHook9( $arg );
+}
+interface Hook10Hook {
+	public function onHook10( $arg );
+}
+interface Hook11Hook {
+	public function onHook11( $arg );
+}
+interface Hook12Hook {
+	public function onHook12( $arg );
+}
+interface Hook13Hook {
+	public function onHook13( $arg );
+}
+interface Hook14Hook {
+	public function onHook14( $arg );
+}
+interface Hook15Hook {
+	public function onHook15( $arg );
+}
+interface Hook16Hook {
+	public function onHook16( $arg );
+}
+
+class HookRunner implements Hook1Hook, Hook2Hook, Hook3Hook, Hook4Hook,
+	Hook5Hook, Hook6Hook, Hook7Hook, Hook8Hook, Hook9Hook, Hook10Hook,
+	Hook11Hook, Hook12Hook, Hook13Hook, Hook14Hook, Hook15Hook, Hook16Hook {
+	public function onHook1( $arg ) {
+	}
+
+	public function onHook2( $arg ) {
+	}
+
+	public function onHook3( $arg ) {
+	}
+
+	public function onHook4( $arg ) {
+	}
+
+	public function onHook5( $arg ) {
+	}
+
+	public function onHook6( $arg ) {
+	}
+
+	public function onHook7( $arg ) {
+	}
+
+	public function onHook8( $arg ) {
+	}
+
+	public function onHook9( $arg ) {
+	}
+
+	public function onHook10( $arg ) {
+	}
+
+	public function onHook11( $arg ) {
+	}
+
+	public function onHook12( $arg ) {
+	}
+
+	public function onHook13( $arg ) {
+	}
+
+	public function onHook14( $arg ) {
+	}
+
+	public function onHook15( $arg ) {
+	}
+
+	public function onHook16( $arg ) {
+	}
+}