+++ tests/integration/hookregistration/test.php
@@ -1,13 +1,13 @@
 <?php
+
 // Test for all kinds of hook registration
-class Hooks {
-	public static function run( $event, array $args = [] ) {
-	}
-}
+
+namespace HookRegistration;
 
 class HookRegistration {
 	public function register() {
 		global $wgHooks;
+
 		$wgHooks['Hook1'][] = $this;
 		$wgHooks['Hook2'][] = [ $this, 'handler2' ];
 		$thirdHandler = [ $this, 'handler3' ];
@@ -17,15 +17,15 @@ class HookRegistration {
 		};
 		$wgHooks['Hook4'][] = $clos;
 		$wgHooks['Hook5'][] = function ( $x ) { echo $x; };
-		$wgHooks['Hook6'][] = 'globalHandler';
-		$wgHooks['Hook7'][] = 'HookRegistration::staticHandler';
+		$wgHooks['Hook6'][] = '\HookRegistration\functionHandler';
+		$wgHooks['Hook7'][] = '\HookRegistration\HookRegistration::staticHandler';
 		$wgHooks['Hook8'][] = new self;
 		$wgHooks['Hook9'][] = new static;
 		$wgHooks['Hook10'][] = new HookRegistration;
 		$wgHooks['Hook11'][] = [ new HookRegistration, 'onHook11', $_GET['unsafe'] ];
 		$wgHooks['Hook12'][] = [ [ $this, 'myHook12Handler' ] ];
 		$wgHooks['Hook13'][] = [ [ $this, 'myHook13Handler' ], $_GET['unsafe'] ];
-		$wgHooks['Hook14'][] = [ 'HookRegistration::staticHandler2' ];
+		$wgHooks['Hook14'][] = [ '\HookRegistration\HookRegistration::staticHandler2' ];
 		$methodName = 'hook15Handler';
 		$wgHooks['Hook15'][] = [ $this, $methodName ];
 		$thisClass = __CLASS__;
@@ -75,25 +75,25 @@ class HookRegistration {
 	}
 }
 
-function globalHandler( $x ) {
+function functionHandler( $x ) {
 	echo $x;
 }
 // These are all unsafe.
-Hooks::run( 'Hook1', [ $_GET['foo'] ] );
-Hooks::run( 'Hook2', [ $_GET['foo'] ] );
-Hooks::run( 'Hook3', [ $_GET['foo'] ] );
-Hooks::run( 'Hook4', [ $_GET['foo'] ] );
-Hooks::run( 'Hook5', [ $_GET['foo'] ] );
-Hooks::run( 'Hook6', [ $_GET['foo'] ] );
-Hooks::run( 'Hook7', [ $_GET['foo'] ] );
-Hooks::run( 'Hook8', [ $_GET['foo'] ] );
-Hooks::run( 'Hook9', [ $_GET['foo'] ] );
-Hooks::run( 'Hook10', [ $_GET['foo'] ] );
-Hooks::run( 'Hook11', [ 'safe' ] ); // TODO Unsafe because of the extra arg registered in __construct
-Hooks::run( 'Hook11', [ $_GET['foo'] ] );
-Hooks::run( 'Hook12', [ $_GET['foo'] ] );
-Hooks::run( 'Hook13', [ 'safe' ] ); // TODO Unsafe because of the extra arg registered in __construct
-Hooks::run( 'Hook13', [ $_GET['foo'] ] );
-Hooks::run( 'Hook14', [ $_GET['foo'] ] );
-Hooks::run( 'Hook15', [ $_GET['foo'] ] );
-Hooks::run( 'Hook16', [ $_GET['foo'] ] );
+( new HookRunner )->onHook1( $_GET['foo'] );
+( new HookRunner )->onHook2( $_GET['foo'] );
+( new HookRunner )->onHook3( $_GET['foo'] );
+( new HookRunner )->onHook4( $_GET['foo'] );
+( new HookRunner )->onHook5( $_GET['foo'] );
+( new HookRunner )->onHook6( $_GET['foo'] );
+( new HookRunner )->onHook7( $_GET['foo'] );
+( new HookRunner )->onHook8( $_GET['foo'] );
+( new HookRunner )->onHook9( $_GET['foo'] );
+( new HookRunner )->onHook10( $_GET['foo'] );
+( new HookRunner )->onHook11( 'safe' ); // TODO Unsafe because of the extra arg registered in __construct
+( new HookRunner )->onHook11( $_GET['foo'] );
+( new HookRunner )->onHook12( $_GET['foo'] );
+( new HookRunner )->onHook13( 'safe' ); // TODO Unsafe because of the extra arg registered in __construct
+( new HookRunner )->onHook13( $_GET['foo'] );
+( new HookRunner )->onHook14( $_GET['foo'] );
+( new HookRunner )->onHook15( $_GET['foo'] );
+( new HookRunner )->onHook16( $_GET['foo'] );