+++ tests/integration/registerhook/HookRunner.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace RegisterHook;
+
+interface SomethingHook {
+	public function onSomething( &$arg1, &$arg2 );
+}
+
+class HookRunner implements SomethingHook {
+	public function onSomething( &$arg1, &$arg2 ) {
+	}
+}