+++ tests/integration/registerhook/Hooks.php
@@ -1,7 +1,0 @@
-<?php
-$wgHooks = [];
-
-class Hooks {
-	public static function run( $name, $args ) {
-	}
-}