+++ tests/integration/skinjson/HookRunner.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace SkinJson;
+
+interface Hook1Hook {
+	public function onHook1( $x, $y, $z );
+}
+interface Hook2Hook {
+	public function onHook2( &$arg );
+}
+
+class HookRunner implements Hook1Hook, Hook2Hook {
+	public function onHook1( $x, $y, $z ) {
+	}
+
+	public function onHook2( &$arg ) {
+	}
+}